Split the SunCalc out
